Make xz a usable package on things, and fix nginx's logrotate to use it.
xz was/is a virtual resource, that's never realised, nor liable to work on
a bunch of platforms. This fixes that, realises it in at least one place,
gives the programme path for the OSes it runs on... This means nginx can
actually it in it's logrotate too, as FreeBSD has a different path to
say Debian.