(#15259) Fix extensions on compressed logs.
logrotate can't work out xz files on it's own, so hold it's hand like the
dumb shit it is.

Manually tell the poor poor programme what the extension is going to be,
as if computers aren't capable of doing man's bidding for them.

Give it an uncompress command in logrotate too, for whatever reason it
might possibly need that.