Redirect HTTP->HTTPS with $host variable
Use `$host` instead of `$server_name` when redirecting from HTTP to HTTPS.
This means that the client will continue using the same hostname that they
came in on, rather than being redirected out to the canonical address, which
is the first value of the `server_name` directive. This is useful for server
aliases (additional args to `server_name`).