Don't use named upstreams for single backend proxies
There's no advantage to using named upstreams in this way. In fact, the
specification of max_fails and fail_timeout with a single backend could
potentially result in the single backend being marked as down and no
requests would be forwarded.

In addition, named upstreams declared in this way (inline in the vhost
template) cause problems when multiple vhosts want to proxy to the same
upstream host, even on different ports, as upstreams cannot be
redeclared with the same name.